chore(releases): moving bundles/components to releases/components
