+++ packages/sanity/src/core/releases/components/dialog/__tests__/ReleaseIconEditorPicker.test.tsx
