+++ packages/sanity/src/core/releases/tool/detail/__tests__/ReleaseReview.test.tsx
@@ -131,7 +131,7 @@ jest.mock('sanity/router', () => ({
   }),
 }))
 
-jest.mock('../../../../bundles/components/ReleasesMenu', () => ({
+jest.mock('../../../components/ReleasesMenu', () => ({
   ReleasesMenu: () => <div>ReleasesMenu</div>,
 }))
