+++ packages/sanity/src/core/releases/tool/overview/ReleasesOverview.tsx
@@ -4,9 +4,9 @@ import {isBefore} from 'date-fns'
 import {type MouseEventHandler, useCallback, useEffect, useMemo, useRef, useState} from 'react'
 
 import {Button, Button as StudioButton} from '../../../../ui-components'
-import {ReleaseDetailsDialog} from '../../../bundles/components/dialog/ReleaseDetailsDialog'
 import {useTranslation} from '../../../i18n'
 import {type BundleDocument, useBundles} from '../../../store'
+import {ReleaseDetailsDialog} from '../../components/dialog/ReleaseDetailsDialog'
 import {releasesLocaleNamespace} from '../../i18n'
 import {containsBundles} from '../../types/bundle'
 import {ReleaseMenuButton} from '../components/ReleaseMenuButton/ReleaseMenuButton'
