+++ packages/sanity/src/core/releases/tool/overview/ReleasesOverviewColumnDefs.tsx
@@ -3,8 +3,8 @@ import {type TFunction, Translate, useTranslation} from 'sanity'
 import {useRouter} from 'sanity/router'
 
 import {Tooltip} from '../../../../ui-components'
-import {ReleaseBadge} from '../../../bundles'
 import {RelativeTime, UserAvatar} from '../../../components'
+import {ReleaseBadge} from '../../components/ReleaseBadge'
 import {releasesLocaleNamespace} from '../../i18n'
 import {type TableRowProps} from '../components/Table/Table'
 import {Headers} from '../components/Table/TableHeader'
