+++ packages/sanity/src/core/studio/components/navbar/perspective/GlobalPerspectiveMenu.tsx
@@ -5,10 +5,10 @@ import {useCallback, useMemo, useState} from 'react'
 import {useTranslation} from 'sanity'
 
 import {MenuItem} from '../../../../../ui-components'
-import {ReleaseDetailsDialog} from '../../../../bundles/components/dialog/ReleaseDetailsDialog'
-import {ReleaseBadge} from '../../../../bundles/components/ReleaseBadge'
-import {ReleasesMenu} from '../../../../bundles/components/ReleasesMenu'
 import {usePerspective} from '../../../../bundles/hooks/usePerspective'
+import {ReleaseDetailsDialog} from '../../../../releases/components/dialog/ReleaseDetailsDialog'
+import {ReleaseBadge} from '../../../../releases/components/ReleaseBadge'
+import {ReleasesMenu} from '../../../../releases/components/ReleasesMenu'
 import {useBundles} from '../../../../store/bundles'
 
 export function GlobalPerspectiveMenu(): JSX.Element {
